math tutor
Problem and Requirements
Write a program that can be used as a math tutor for an elementary student. 
The program should display two random integer numbers.
The program should wait for the students to enter the answer.
If the answer is correct, a message of congratulations should be printed.
If the answer is incorrect, a message should be printed showing the correct answer.
The program displays a menu allowing the user to select an addition, subtraction, multiplication, or division problem.
The final selection on the menu should let the user quit the program.
After the user has finished the math problem, the program should display the menu again. 
This process is repeated until the user chooses to quit the program.
Input Validation: If the user select an item not on the menu, display an error message and display the menu again.*/