two d rray
Write a program that creates a two-dimensional array initialized with test data.
The program should have the following functions:

getTotal - This function should accept a two-dimensional array as its argument and return the total of all the values in the array.

getAverage - This function should accept a two-dimensional array as its argument and return the average of values in the array.

getRowTotal - This function should accept a two-dimensional array as its first argument and an integer as its second argument.
The second argument should be the subscript of a row in the array.
The function should return the total of the values in the specified row.

getColumnTotal - This function should accept a two-dimensional array as its first argument and an integer as its second argument.
The second argument should be the subscript of a column in the array.
The function should return the total of the values in the specified column.

getHighestInRow - This function should accept a two-dimensional array as its first argument and an integer as its second argument.
The second argument should be the subscript of a row in the array. 
The function should return the highest value in the specified row in the array.

getLowestInRow - This function should accept a two-dimensional array as its first argument and an integer as its second argument.
The second argument should be the subscript of a row in the array.
The function should return the lowest value in the specified row in the array.