circle class
Write a program the demonstrate the Circle class performs the following tasks:
Create a Circle object called C1 with a radius equal to zero.
Create a Circle object called C2 with a radius equal to 5.0
Input the radius for the circle C1 from the keyboard. Enter a value of 3.0 for the radius of the circle C1.
Display the area of the circle C2 by calling the function getArea.
Display the diameter of the circle C2 by calling the function getDiameter.
Display the radius and area of the circle C1 by calling the function displayCircle.
Change the radius of the circle C1 to 10.0.
Display the circumference of the circle C1 by calling the function getCircumference.